+++ geo_activity_playground/core/activities.py
@@ -48,9 +48,6 @@ class ActivityRepository:
         return len(self.meta)
 
     def add_activity(self, activity_meta: ActivityMeta) -> None:
-        assert not self.has_activity(
-            activity_meta["id"]
-        ), f"Trying to add the following activity which already exists: {activity_meta}"
         self._loose_activities.append(activity_meta)
 
     def commit(self) -> None:
@@ -59,7 +56,15 @@ class ActivityRepository:
                 f"Adding {len(self._loose_activities)} activities to the repository …"
             )
             new_df = pd.DataFrame(self._loose_activities)
-            self.meta = pd.concat([self.meta, new_df])
+            if len(self.meta):
+                new_ids_set = set(new_df["id"])
+                is_kept = [
+                    activity_id not in new_ids_set for activity_id in self.meta["id"]
+                ]
+                old_df = self.meta.loc[is_kept]
+            else:
+                old_df = self.meta
+            self.meta = pd.concat([old_df, new_df])
             assert pd.api.types.is_dtype_equal(
                 self.meta["start"].dtype, "datetime64[ns, UTC]"
             ), self.meta["start"].dtype