+++ docs/reference/changelog.md
@@ -9,6 +9,7 @@ This is the pre-release series. Things haven't settled yet, so each minor versio
 ### Version 0.20.0
 
 - Import altitude information from GPX files if available. [GH-91](https://github.com/martin-ueding/geo-activity-playground/pull/91)
+- Keep identity of activities based on hash of the file content, not the path. This allows to rename activities and just update their metadata, without having duplicates. [GH-92](https://github.com/martin-ueding/geo-activity-playground/issues/92).
 
 ### Version 0.19.1